Successfully parsing [Component].
